# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ANVISA-RE Monitor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export interface Resolucao {
  id: string;
  numero_re: string;
  data_publicacao: string;
  data_vigencia_inicio: string;
  /** @nullable */
  data_vigencia_fim?: string | null;
  /** vigente | revogada | encerrada | em_analise */
  status: string;
  tipo_produto: string;
  tipo_acao: string[];
  nome_produto: string;
  /** @nullable */
  principio_ativo?: string | null;
  /** @nullable */
  fabricante_nome?: string | null;
  /** @nullable */
  fabricante_cnpj?: string | null;
  /** @nullable */
  numero_registro_anvisa?: string | null;
  ementa: string;
  /** @nullable */
  link_documento_oficial?: string | null;
  /** @nullable */
  arquivo_pdf_path?: string | null;
  /** @nullable */
  criado_por?: string | null;
  created_at: string;
  updated_at: string;
  /** @nullable */
  deleted_at?: string | null;
}

export interface ResolucaoResumida {
  id: string;
  numero_re: string;
  nome_produto: string;
  status: string;
  tipo_produto: string;
  data_publicacao: string;
  /** @nullable */
  fabricante_nome?: string | null;
  updated_at: string;
}

export interface PaginatedResolucoes {
  data: Resolucao[];
  total: number;
  page: number;
  limit: number;
  total_pages: number;
}

export interface CreateResolucaoBody {
  numero_re: string;
  data_publicacao: string;
  data_vigencia_inicio: string;
  /** @nullable */
  data_vigencia_fim?: string | null;
  status: string;
  tipo_produto: string;
  tipo_acao: string[];
  nome_produto: string;
  /** @nullable */
  principio_ativo?: string | null;
  /** @nullable */
  fabricante_nome?: string | null;
  /** @nullable */
  fabricante_cnpj?: string | null;
  /** @nullable */
  numero_registro_anvisa?: string | null;
  ementa: string;
  /** @nullable */
  link_documento_oficial?: string | null;
  /** @nullable */
  arquivo_pdf_path?: string | null;
}

export interface UpdateResolucaoBody {
  /** @nullable */
  data_vigencia_fim?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  tipo_produto?: string | null;
  /** @nullable */
  tipo_acao?: string[] | null;
  /** @nullable */
  nome_produto?: string | null;
  /** @nullable */
  principio_ativo?: string | null;
  /** @nullable */
  fabricante_nome?: string | null;
  /** @nullable */
  fabricante_cnpj?: string | null;
  /** @nullable */
  numero_registro_anvisa?: string | null;
  /** @nullable */
  ementa?: string | null;
  /** @nullable */
  link_documento_oficial?: string | null;
  /** @nullable */
  arquivo_pdf_path?: string | null;
  /** @nullable */
  justificativa_alteracao?: string | null;
}

/**
 * @nullable
 */
export type HistoricoEntryDadosAnteriores = { [key: string]: unknown } | null;

/**
 * @nullable
 */
export type HistoricoEntryDadosNovos = { [key: string]: unknown } | null;

export interface HistoricoEntry {
  id: string;
  resolucao_id: string;
  /** @nullable */
  campo_alterado?: string | null;
  /** @nullable */
  valor_anterior?: string | null;
  /** @nullable */
  valor_novo?: string | null;
  /** @nullable */
  justificativa?: string | null;
  /** @nullable */
  alterado_por?: string | null;
  /** @nullable */
  alterado_por_nome?: string | null;
  fonte: string;
  created_at: string;
  /** @nullable */
  dados_anteriores?: HistoricoEntryDadosAnteriores;
  /** @nullable */
  dados_novos?: HistoricoEntryDadosNovos;
}

export interface LoginBody {
  email: string;
  senha: string;
}

export interface Usuario {
  id: string;
  nome: string;
  email: string;
  /** administrador | fiscal | profissional_saude | cidadao */
  perfil: string;
  ativo: boolean;
  created_at: string;
  /** @nullable */
  ultimo_acesso?: string | null;
}

export interface AuthResponse {
  token: string;
  usuario: Usuario;
}

export interface PaginatedUsuarios {
  data: Usuario[];
  total: number;
  page: number;
  limit: number;
  total_pages: number;
}

export interface CreateUsuarioBody {
  nome: string;
  email: string;
  senha: string;
  perfil: string;
}

export interface UpdateUsuarioBody {
  /** @nullable */
  nome?: string | null;
  /** @nullable */
  perfil?: string | null;
  /** @nullable */
  ativo?: boolean | null;
  /** @nullable */
  senha?: string | null;
}

export interface DashboardStats {
  total_resolucoes: number;
  vigentes: number;
  revogadas: number;
  encerradas: number;
  em_analise: number;
  novas_este_mes: number;
  total_usuarios: number;
  ultima_atualizacao: string;
}

export interface CategoriaCount {
  categoria: string;
  total: number;
  vigentes: number;
}

export interface StatusCount {
  status: string;
  total: number;
}

export interface RelatorioMensal {
  ano: number;
  mes: number;
  total_novas: number;
  por_categoria: CategoriaCount[];
  por_status: StatusCount[];
  resolucoes: ResolucaoResumida[];
}

export interface ExportResponse {
  url: string;
  formato: string;
  total_registros: number;
  filename: string;
}

export interface Categoria {
  id: string;
  nome: string;
  /** @nullable */
  descricao?: string | null;
  /** @nullable */
  exemplos?: string | null;
}

export interface TipoAcao {
  id: string;
  nome: string;
  /** @nullable */
  descricao?: string | null;
}

export type ListResolucoesParams = {
  /**
   * Busca full-text (nome do produto, princípio ativo, fabricante, ementa)
   * @nullable
   */
  q?: string | null;
  /**
   * Filtro por status: vigente, revogada, encerrada, em_analise
   * @nullable
   */
  status?: string | null;
  /**
   * Filtro por categoria de produto
   * @nullable
   */
  tipo_produto?: string | null;
  /**
   * Filtro por tipo de ação sanitária
   * @nullable
   */
  tipo_acao?: string | null;
  /**
   * Data de publicação início (ISO 8601: YYYY-MM-DD)
   * @nullable
   */
  data_inicio?: string | null;
  /**
   * Data de publicação fim (ISO 8601: YYYY-MM-DD)
   * @nullable
   */
  data_fim?: string | null;
  /**
   * Número exato da RE
   * @nullable
   */
  numero_re?: string | null;
  /**
   * Número de registro ANVISA
   * @nullable
   */
  numero_registro_anvisa?: string | null;
  /**
   * CNPJ do fabricante
   * @nullable
   */
  fabricante_cnpj?: string | null;
  /**
   * Página (default 1)
   * @nullable
   */
  page?: number | null;
  /**
   * Itens por página (default 20, max 100)
   * @nullable
   */
  limit?: number | null;
  /**
   * Campo de ordenação (default: data_publicacao)
   * @nullable
   */
  order_by?: string | null;
  /**
   * Direção: asc ou desc
   * @nullable
   */
  order_dir?: string | null;
};

export type ExportarResolucoesParams = {
  /**
   * Formato de exportação: csv, xlsx, pdf
   */
  formato: string;
  /**
   * @nullable
   */
  q?: string | null;
  /**
   * @nullable
   */
  status?: string | null;
  /**
   * @nullable
   */
  tipo_produto?: string | null;
  /**
   * @nullable
   */
  tipo_acao?: string | null;
  /**
   * @nullable
   */
  data_inicio?: string | null;
  /**
   * @nullable
   */
  data_fim?: string | null;
};

export type ListUsuariosParams = {
  /**
   * @nullable
   */
  page?: number | null;
  /**
   * @nullable
   */
  limit?: number | null;
  /**
   * @nullable
   */
  q?: string | null;
};

export type GetRecentesResolucoesParams = {
  /**
   * @nullable
   */
  limit?: number | null;
};

export type GetRelatorioMensalParams = {
  /**
   * @nullable
   */
  ano?: number | null;
  /**
   * @nullable
   */
  mes?: number | null;
};