# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ANVISA-RE Monitor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Autenticação e emissão de token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Autenticado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Retorna dados do usuário autenticado
      responses:
        "200":
          description: Dados do usuário atual
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
        "401":
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/dashboard/recentes:
    get:
      operationId: getRecentesResolucoes
      tags: [dashboard]
      summary: Últimas REs cadastradas ou atualizadas
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: REs recentes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResolucaoResumida"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ResolucaoResumida:
      type: object
      properties:
        id:
          type: string
        numero_re:
          type: string
        nome_produto:
          type: string
        status:
          type: string
        tipo_produto:
          type: string
        data_publicacao:
          type: string
        fabricante_nome:
          type: ["string", "null"]
        updated_at:
          type: string
      required:
        - id
        - numero_re
        - nome_produto
        - status
        - tipo_produto
        - data_publicacao
        - updated_at
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        senha:
          type: string
      required:
        - email
        - senha
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        usuario:
          $ref: "#/components/schemas/Usuario"
      required:
        - token
        - usuario
    Usuario:
      type: object
      properties:
        id:
          type: string
        nome:
          type: string
        email:
          type: string
        perfil:
          type: string
          description: "administrador | fiscal | profissional_saude | cidadao"
        ativo:
          type: boolean
        created_at:
          type: string
        ultimo_acesso:
          type: ["string", "null"]
      required:
        - id
        - nome
        - email
        - perfil
        - ativo
        - created_at
...